IoT_Gateway/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
websockets==11.0.3
